+++ small_text/integrations/pytorch/classifiers/kimcnn.py
@@ -274,7 +274,7 @@ class KimCNNClassifier(KimCNNEmbeddingMixin, PytorchClassifier):
             self.initialize_kimcnn_model(sub_train)
 
         check_optimizer_and_scheduler_config(optimizer, scheduler)
-        scheduler = scheduler if scheduler is None else None
+        scheduler = scheduler if scheduler is not None else None
 
         optimizer, scheduler = self._get_optimizer_and_scheduler(optimizer,
                                                                  scheduler,